requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
ansible==2.4.3
asn1crypto==0.22.0        # via cryptography
bcrypt==3.1.3             # via paramiko
certifi==2017.7.27.1      # via requests
cffi==1.11.0              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
click==6.7                # via flask
cryptography==2.0.3       # via paramiko, pyopenssl
enum34==1.1.6
flask==0.12.2
idna==2.6                 # via cryptography, requests
ipaddress==1.0.18         # via cryptography
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
lxml==4.0.0
m2crypto==0.26.0          # via pywbem
markupsafe==1.0           # via jinja2
numpy==1.13.1
pandas==0.19.2
paramiko==2.3.0
ply==3.10
protobuf==3.1.0
py==1.4.34
pyasn1==0.3.5             # via paramiko
pycparser==2.18           # via cffi
pycrypto==2.6.1
pynacl==1.1.2             # via paramiko
pyopenssl==17.3.0         # via pyvmomi
python-dateutil==2.6.1
python-gflags==3.1.1
pytz==2017.2
pyvmomi==6.5.0.2017.5
pywbem==0.10.0
pyyaml==3.12              # via pywbem
requests==2.18.4
scp==0.10.2
six==1.11.0               # via bcrypt, cryptography, protobuf, pynacl, pyopenssl, python-dateutil, pyvmomi, pywbem
typing==3.6.2             # via m2crypto
urllib3==1.22             # via requests
werkzeug==0.12.2          # via flask
